Middleware: jwt codes optimization.

+++ config/config.go
@@ -31,7 +31,7 @@ type Mail struct {
 
 type Auth struct {
 	JWTSecret       string `toml:"jwt_secret"`
-	JWTLifetime     int    `toml:"jwt_lifetime"`
+	JWTLifetime     int64  `toml:"jwt_lifetime"`
 	APISignLifetime int64  `toml:"apisign_lifetime"`
 }
